feat: update Mikvah model types and enhance restaurant service filtering
- Changed data types for `latitude` and `longitude` in the Mikvah model to `Numeric` for better precision.
- Updated the EntityRepositoryV5 to use `name` instead of `title` for sorting options.
- Enhanced the RestaurantServiceV5 to extract and include dynamic ratings and kosher details for improved filtering options.
- Refactored CSRF middleware to remove debug logging for cleaner output.
- Improved logging in the EateryPageClient and ModernFilterPopup components for better debugging of filter options.

+++ backend/database/models.py
@@ -468,8 +468,8 @@ class Mikvah(Base):
     email = Column(String(255))
     
     # Location data
-    latitude = Column(Float)  # Changed from Numeric to Float to match Synagogue
-    longitude = Column(Float)  # Changed from Numeric to Float to match Synagogue
+    latitude = Column(Numeric(10, 8))  # Match actual database type
+    longitude = Column(Numeric(11, 8))  # Match actual database type
     
     # Mikvah details
     mikvah_type = Column(String(100))
